move StateVec to state module and fix some documentation issues
