+++ src/phase_equilibria/phase_diagram_pure.rs
@@ -1,10 +1,10 @@
 use super::{PhaseEquilibrium, SolverOptions};
-use crate::equation_of_state::{EquationOfState, MolarWeight};
+use crate::equation_of_state::EquationOfState;
 use crate::errors::EosResult;
-use crate::state::{Contributions, State};
+use crate::state::{State, StateVec};
 use crate::EosUnit;
 use ndarray::prelude::*;
-use quantity::{QuantityArray1, QuantityScalar};
+use quantity::QuantityScalar;
 use std::rc::Rc;
 
 /// Pure component and binary mixture phase diagrams.
@@ -65,77 +65,5 @@ impl<U: EosUnit, E: EquationOfState> PhaseDiagram<U, E> {
         StateVec {
             states: self.states.iter().map(|s| s.liquid()).collect(),
         }
-    }
-}
-
-/// A list of states for a simple access to properties
-/// of multiple states.
-pub struct StateVec<'a, U, E> {
-    pub states: Vec<&'a State<U, E>>,
-}
-
-impl<'a, U: EosUnit, E: EquationOfState> StateVec<'a, U, E> {
-    pub fn temperature(&self) -> QuantityArray1<U> {
-        QuantityArray1::from_shape_fn(self.states.len(), |i| self.states[i].temperature)
-    }
-
-    pub fn pressure(&self) -> QuantityArray1<U> {
-        QuantityArray1::from_shape_fn(self.states.len(), |i| {
-            self.states[i].pressure(Contributions::Total)
-        })
-    }
-
-    pub fn compressibility(&self) -> Array1<f64> {
-        Array1::from_shape_fn(self.states.len(), |i| {
-            self.states[i].compressibility(Contributions::Total)
-        })
-    }
-
-    pub fn density(&self) -> QuantityArray1<U> {
-        QuantityArray1::from_shape_fn(self.states.len(), |i| self.states[i].density)
-    }
-
-    pub fn molefracs(&self) -> Array2<f64> {
-        Array2::from_shape_fn(
-            (self.states.len(), self.states[0].eos.components()),
-            |(i, j)| self.states[i].molefracs[j],
-        )
-    }
-
-    pub fn molar_enthalpy(&self) -> QuantityArray1<U> {
-        QuantityArray1::from_shape_fn(self.states.len(), |i| {
-            self.states[i].molar_enthalpy(Contributions::Total)
-        })
-    }
-
-    pub fn molar_entropy(&self) -> QuantityArray1<U> {
-        QuantityArray1::from_shape_fn(self.states.len(), |i| {
-            self.states[i].molar_entropy(Contributions::Total)
-        })
-    }
-}
-
-impl<'a, U: EosUnit, E: EquationOfState + MolarWeight<U>> StateVec<'a, U, E> {
-    pub fn mass_density(&self) -> QuantityArray1<U> {
-        QuantityArray1::from_shape_fn(self.states.len(), |i| self.states[i].mass_density())
-    }
-
-    pub fn massfracs(&self) -> Array2<f64> {
-        Array2::from_shape_fn(
-            (self.states.len(), self.states[0].eos.components()),
-            |(i, j)| self.states[i].massfracs()[j],
-        )
-    }
-
-    pub fn specific_enthalpy(&self) -> QuantityArray1<U> {
-        QuantityArray1::from_shape_fn(self.states.len(), |i| {
-            self.states[i].specific_enthalpy(Contributions::Total)
-        })
-    }
-
-    pub fn specific_entropy(&self) -> QuantityArray1<U> {
-        QuantityArray1::from_shape_fn(self.states.len(), |i| {
-            self.states[i].specific_entropy(Contributions::Total)
-        })
     }
 }
